+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Changelog
 ---
 
+> Saturday 9th April 2022
+
+- Added thumbnail.js command
+
 > Friday 8th April 2022
 
 - Uploaded existing commands from GymRat